Update gulp to 4.x
